Add support for EL 7
This is a mostly straightforward port of the Ubuntu task to
RedHat/CentOS 7.2 and a "wrapper" as main.yml which selects
the task that should run depending on the OS family.

Storage setup on EL 7 is not performed, so the default Docker
backend for EL (devicemapper on loopback) will be selected.